Add double-headed arrows to GrowArea graphics
GrowAreas will draw a horizontal arrow if they have any
xGrowth and a vertical arrow if they have any yGrowth.
The arrow design is rudimentary and will hopefully be
polished more in the future. Right now, it is the minimum
arrow to help the player understand the purpose of GrowAreas.