Fix bug with collision direction
If a MovingRectangle (A) was fudged on top of a wall which
another MovingRectangle (B) was sitting on, and A was
entirely outside of B's bounds on the previous frame, A
would push B above it instead of properly pushing it to the
side.

This fixes that bug by taking the collision direction to
be in the x direction if other (B in the example above)
is above rect (A).